# Run `test_search_model_versions` faster